Move defaults into separate submodule
