+++ src/wily/defaults.py
@@ -1,0 +1,18 @@
+"""Default values."""
+
+""" Default table style in console. See tabulate docs for more. """
+DEFAULT_GRID_STYLE = "fancy_grid"
+
+""" The name of the default archiver """
+DEFAULT_ARCHIVER = "git"
+
+""" The default configuration file name """
+DEFAULT_CONFIG_PATH = "wily.cfg"
+
+""" The default section name in the config """
+DEFAULT_CONFIG_SECTION = "wily"
+
+""" The default maximum number of revisions to archiver """
+DEFAULT_MAX_REVISIONS = 50
+
+DEFAULT_PATH = "."
