+++ src/wily/helper/__init__.py
@@ -1,9 +1,15 @@
 """Helper package for wily."""
+import hashlib
+import logging
+import pathlib
 import shutil
 import sys
+from functools import lru_cache
 from typing import Sized
 
-from wily.config import DEFAULT_GRID_STYLE
+from wily.defaults import DEFAULT_GRID_STYLE
+
+logger = logging.getLogger(__name__)
 
 
 def get_maxcolwidth(headers: Sized, wrap=True):
@@ -28,3 +34,22 @@ def get_style(style: str = DEFAULT_GRID_STYLE):
         if sys.stdout.encoding.lower() not in ("utf-8", "utf8"):
             style = "grid"
     return style
+
+
+@lru_cache(maxsize=128)
+def generate_cache_path(path):
+    """
+    Generate a reusable path to cache results.
+
+    Will use the --path of the target and hash into
+    a 9-character directory within the HOME folder.
+
+    :return: The cache path
+    :rtype: ``str``
+    """
+    logger.debug(f"Generating cache for {path}")
+    sha = hashlib.sha1(str(path).encode()).hexdigest()[:9]
+    HOME = pathlib.Path.home()
+    cache_path = str(HOME / ".wily" / sha)
+    logger.debug(f"Cache path is {cache_path}")
+    return cache_path
